+++ Maya/Exporter/BabylonExporter.GLTFExporter.cs
@@ -25,7 +25,7 @@ namespace Maya2Babylon
 
         public void ExportGltf(BabylonScene babylonScene, string outputDirectory, string outputFileName, bool generateBinary)
         {
-            RaiseMessage("GLTFExporter | Exportation started", Color.Blue);
+            RaiseMessage("GLTFExporter | Export started", Color.Blue);
 
             // Force output file extension to be gltf
             outputFileName = Path.ChangeExtension(outputFileName, "gltf");
@@ -47,8 +47,7 @@ namespace Maya2Babylon
             gltf.asset = new GLTFAsset
             {
                 version = "2.0",
-                generator = $"babylon.js glTF exporter for maya 2018 v{exporterVersion}",
-                copyright = "2017 (c) BabylonJS"
+                generator = $"babylon.js glTF exporter for maya 2018 v{exporterVersion}"
                 // no minVersion
             };
 
@@ -310,7 +309,11 @@ namespace Maya2Babylon
                 }
                 else if (type == typeof(BabylonLight) || type.IsSubclassOf(typeof(BabylonLight)))
                 {
-                    ExportLight(ref gltfNode, babylonNode as BabylonLight, gltf, gltfParentNode, babylonScene);
+                    if (_exportKHRLightsPunctual)
+                    {
+                        RaiseMessage("exporting light", 1);
+                        ExportLight(ref gltfNode, babylonNode as BabylonLight, gltf, gltfParentNode, babylonScene);
+                    }
                 }
                 else
                 {